refactoring. drop unused algorithms, move some stuff around
